Move CC scheme classes to sample file.
